pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rags"
version = "0.1.0"
description = "Retrieval-Augmented Generation pipeline for semantic search and response generation."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "RAG Pipeline Team" },
]
dependencies = [
    "streamlit==1.28.1",
    "python-dotenv==1.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "requests>=2.31.0",
    "accelerate>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
]

[project.urls]
homepage = "https://github.com/"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
src = ["src", "tests"]
line-length = 100
target-version = "py38"
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "build",
    "dist",
    "data",
    "models",
]

[tool.ruff.lint]
select = ["E", "F", "W", "B", "I", "UP"]
fixable = ["ALL"]
preview = false

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

